BREAKING CHANGE: xWebSiteDefaults: Align to best practices
This change brings the modules xWebSiteDefaults resource into
alignment with best practices for single instance resources.

When defining a resource that should only be allowed to be defined once
in any configuration, there are now best practices for how to implement
that restriction. There should be a single key property and its name
should be 'IsSingleInstance' with a value of 'Yes'. This change
implements that best practice.

Fixes #457

[x] Added an entry to the change log under the Unreleased section of the CHANGELOG.md.
Entry should say what was changed and how that affects users (if applicable), and
reference the issue being resolved (if applicable).
[x] Resource documentation added/updated in README.md.
[x] Resource parameter descriptions added/updated in README.md, schema.mof
and comment-based help.
[ ] Comment-based help added/updated.
[ ] Localization strings added/updated in all localization files as appropriate.
[x] Examples appropriately added/updated.
[ ] Unit tests added/updated. See DSC Resource Testing Guidelines.
[x] Integration tests added/updated (where possible). See DSC Resource Testing Guidelines.
[x] New/changed code adheres to DSC Resource Style Guidelines and Best Practices.